feat: add status_data blueprint and implement status data endpoint
